requirements.txt
discord.py==2.3.1
aiohttp==3.8.3
DateTime==4.4
python-dotenv==0.20.0
typing_extensions==4.2.0